backend/ml_services/reid_trainer.py
# backend/ml_services/reid_trainer.py

import os
import json
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import Dataset, DataLoader
from torchvision import transforms
# from torchvision.datasets import ImageFolder # 假设你的数据能组织成 ImageFolder 结构，如果用不到可以不导入
from typing import List, Optional, Tuple, Dict, Any
from PIL import Image
import numpy as np
import logging
import io # 用于处理图像字节流，如果需要
from sqlalchemy.orm import Session # 新增导入 Session

# 导入你的 schema 和 crud
from .. import schemas
from .. import crud
from ..database_conn import SessionLocal
from backend.config import settings

logger = logging.getLogger(__name__)

# --- 1. 数据集定义 ---
class ReIDDataset(Dataset):
    def __init__(self, persons_data: List[Dict[str, Any]], transform=None): # 修改类型提示
        self.persons_data = persons_data
        self.transform = transform
        self.image_paths = []
        self.labels = []
        self.person_id_to_label = {} # 用于将 person.id 映射到整数标签

        current_label = 0
        for person in persons_data:
            relative_crop_path = person.get('crop_image_path')

            if relative_crop_path is None:
                logger.warning(f"人物 {person.get('uuid')} 没有有效的裁剪图片路径 (crop_image_path)，跳过此人物的数据加载。")
                continue

            # 确保 person.crop_image_path 是相对于 DATABASE_CROPS_DIR 的相对路径
            # 移除可能重复的前缀
            if relative_crop_path.startswith("database/crops/"):
                relative_crop_path = relative_crop_path.replace("database/crops/", "", 1)
            elif relative_crop_path.startswith("database\\crops\\"): # 针对Windows路径
                relative_crop_path = relative_crop_path.replace("database\\crops\\", "", 1)
            
            crop_image_full_path = os.path.join(settings.DATABASE_CROPS_DIR, relative_crop_path)
            if not os.path.exists(crop_image_full_path):
                logger.warning(f"ReIDDataset: 裁剪图文件不存在，跳过: {crop_image_full_path}")
                continue
            
            # TODO: **重要** - 这里需要更复杂的逻辑来处理纠正信息，生成有意义的标签。
            # 如果有合并操作，需要将多个 person.uuid 映射到同一个标签。
            # 如果是重新标注，则返回新的标注ID。
            # 否则，返回 person.uuid 对应的 ID。
            # 建议维护一个从 person_uuid 到全局唯一整数ID的映射。
            person_uuid = person.get('uuid') # 从字典获取 uuid
            if person_uuid and person_uuid not in self.person_id_to_label: # 确保 uuid 存在
                self.person_id_to_label[person_uuid] = current_label
                current_label += 1
            label = self.person_id_to_label[person_uuid]

            self.image_paths.append(crop_image_full_path)
            self.labels.append(label)

        logger.info(f"ReIDDataset: 已准备 {len(self.image_paths)} 个训练样本，共 {len(self.person_id_to_label)} 个独立身份。")

    def __len__(self):
        return len(self.image_paths)

    def __getitem__(self, idx):
        img_path = self.image_paths[idx]
        label = self.labels[idx]
        image = Image.open(img_path).convert("RGB") # 确保是RGB
        
        if self.transform:
            image = self.transform(image)
        
        return image, label

# --- 2. 模型定义 (示例：简化版模型，你需要替换为真正的 Re-ID 模型，例如 OSNet) ---
# 你可能需要从外部库导入 OSNet，或者在这里定义其结构。
# 例如：
# from torchvision.models import resnet50
# class OSNet(nn.Module):
#     def __init__(self, num_classes):
#         super(OSNet, self).__init__()
#         # ... OSNet 结构 ...

class DummyReIDModel(nn.Module):
    def __init__(self, num_classes: int, feature_dim: int = 512):
        super(DummyReIDModel, self).__init__()
        # 这是一个非常简化的模型，实际的 Re-ID 模型会复杂得多
        # 请替换为你的 OSNet 或其他 Re-ID 模型实现
        self.base_model = nn.Sequential(
            nn.Conv2d(3, 32, kernel_size=3, stride=1, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(kernel_size=2, stride=2),
            nn.Conv2d(32, 64, kernel_size=3, stride=1, padding=1),
            nn.ReLU(),
            nn.MaxPool2d(kernel_size=2, stride=2),
            nn.Flatten()
        )
        # 调整这里以匹配你的特征维度和分类头
        # 注意：这里的输入尺寸假设为 128x256，经过两次MaxPool2d变为 32x64，所以特征尺寸是 64 * 64 * 32
        self.fc = nn.Linear(64 * 64 * 32, feature_dim) # 修正：根据实际输出尺寸计算
        self.classifier = nn.Linear(feature_dim, num_classes) # 分类头

    def forward(self, x):
        x = self.base_model(x)
        features = self.fc(x)
        # 归一化特征，这对于 Re-ID 很重要
        features = features / torch.norm(features, p=2, dim=1, keepdim=True).clamp(min=1e-12)
        logits = self.classifier(features) # 如果模型有分类头
        return features, logits # 返回特征和分类 logits


# --- 3. 训练函数 ---
def train_reid_model_pytorch(
    persons_for_retrain: List[schemas.Person],
    new_model_path: str,
    db: Session,
    celery_task = None,
    epochs: int = 10,
    batch_size: int = 32,
    learning_rate: float = 0.001,
    feature_dim: int = 512,
    global_progress_start: int = 0, # 新增参数：全局进度起始百分比
    global_progress_end: int = 100 # 新增参数：全局进度结束百分比
) -> bool:
    logger.info("开始 PyTorch 模型训练。")

    # 确保 CUDA 可用
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    logger.info(f"训练设备: {device}")

    # 图像预处理
    transform = transforms.Compose([
        transforms.Resize((settings.REID_INPUT_HEIGHT, settings.REID_INPUT_WIDTH)), # 使用 settings 中的尺寸
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
    ])

    # 创建数据集和数据加载器
    dataset = ReIDDataset(persons_for_retrain, transform=transform)
    if len(dataset) == 0:
        logger.warning("没有有效的训练数据，跳过训练。")
        return False

    # 获取类别数量
    num_classes = len(dataset.person_id_to_label)
    if num_classes < 2: # 至少需要两个类别才能进行有意义的分类训练
        logger.warning(f"类别数量不足 ({num_classes})，无法进行分类训练。")
        # 对于 Re-ID，即使类别少，也可以尝试度量学习，但分类头可能无意义
        # 在这种情况下，你可以选择跳过分类损失或使用其他策略
        return False

    dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=0) # 将 num_workers 设置为 0

    # 初始化模型
    # TODO: **重要** - 将 DummyReIDModel 替换为您的实际 Re-ID 模型（例如 OSNet）
    # 您可能需要先加载预训练的权重，然后根据您的需求进行微调。
    # 例如：model = OSNet(num_classes=num_classes) 
    # model.load_state_dict(torch.load('path/to/your/pretrained_osnet.pth'))
    # model.classifier = nn.Linear(feature_dim, num_classes) # 修改分类头以匹配新的类别数
    model = DummyReIDModel(num_classes=num_classes, feature_dim=feature_dim).to(device)

    # 定义损失函数和优化器
    # TODO: **重要** - Re-ID 通常结合分类损失（如交叉熵）和度量学习损失（如三元组损失、对比损失）
    # criterion_cls = nn.CrossEntropyLoss()
    # criterion_triplet = TripletLoss(...) # 你需要实现 TripletLoss 或从现有库中导入

    # 这里只用分类损失作为示例
    criterion = nn.CrossEntropyLoss()
    optimizer = optim.Adam(model.parameters(), lr=learning_rate)

    # 训练循环
    for epoch in range(epochs):
        model.train() # 设置为训练模式
        running_loss = 0.0
        for batch_idx, (images, labels) in enumerate(dataloader):
            images = images.to(device)
            labels = labels.to(device)

            optimizer.zero_grad()
            features, logits = model(images) # 假设模型返回特征和 logits
            
            # TODO: 结合分类损失和度量学习损失
            loss = criterion(logits, labels) # 仅使用分类损失作为示例

            loss.backward()
            optimizer.step()

            running_loss += loss.item()

            if celery_task:
                # 更新 Celery 任务进度
                local_progress = int(((epoch * len(dataloader) + batch_idx + 1) / (epochs * len(dataloader))) * 100)
                # 将本地进度映射到全局进度范围
                progress = global_progress_start + int((local_progress / 100) * (global_progress_end - global_progress_start))
                celery_task.update_state(state='PROGRESS', meta={'progress': progress, 'message': f'Training epoch {epoch+1}/{epochs} (Local: {local_progress}%)'})

        logger.info(f"Epoch {epoch+1}/{epochs}, Loss: {running_loss / len(dataloader):.4f}")

    logger.info("PyTorch 模型训练完成。")

    # --- 4. 模型保存为 ONNX 格式 ---
    # 确保模型处于评估模式，这样像 BatchNorm、Dropout 这样的层会固定住
    model.eval() 
    
    # 定义一个 dummy input，这对于 ONNX 导出是必需的
    # 尺寸应与您的模型预期输入尺寸匹配 (batch_size, channels, height, width)
    dummy_input = torch.randn(1, 3, settings.REID_INPUT_HEIGHT, settings.REID_INPUT_WIDTH).to(device) 

    try:
        torch.onnx.export(
            model,
            dummy_input,
            new_model_path,
            export_params=True,
            opset_version=11, # 根据您的 PyTorch 和 ONNX Runtime 版本调整，通常 11 或更高
            do_constant_folding=True,
            input_names=['input'],
            output_names=['features', 'logits'], # 假设模型输出特征和 logits，根据实际模型输出调整
            dynamic_axes={'input': {0: 'batch_size'}, 'features': {0: 'batch_size'}, 'logits': {0: 'batch_size'}} # 允许动态 batch size
        )
        logger.info(f"模型已成功保存为 ONNX 格式: {new_model_path}")
        return True
    except Exception as e:
        logger.error(f"将 PyTorch 模型导出为 ONNX 失败: {e}", exc_info=True)
        return False

# --- 辅助函数：将 ONNX 模型加载为 PyTorch 模型（可选，如果您需要加载 ONNX 后再进行微调）---
# 这通常需要 ONNX Runtime 和 ONNX 的 Python API
# 或者使用 ONNX-PyTorch 等工具，但通常直接从 PyTorch 模型开始训练更方便。
# import onnx
# from onnx_pytorch import code_gen

# def load_onnx_to_pytorch(onnx_model_path: str) -> nn.Module:
#     model = onnx.load(onnx_model_path)
#     pytorch_model_code = code_gen.gen_code(model)
#     # 然后你需要执行这段代码来实例化 PyTorch 模型
#     # 这是一个高级且可能复杂的操作，通常不建议直接在运行时进行。
#     # 更好的方法是维护模型的 PyTorch 源代码。
#     return your_pytorch_model_instance 