requirements.txt
uvicorn
SQLAlchemy
python-multipart
opencv-python
numpy
scikit-image
scipy
imgaug
requests
pillow
PyYAML
loguru
matplotlib
torchvision
torchaudio
torch
fastapi
Jinja2
aiofiles
python-jose[cryptography]
passlib[bcrypt==3.2.0]
python-multpart
httpx
celery
redis
pytz
faiss-cpu